requirements.txt
django>=2.2.13
PyYAML>=5.3.1